requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
pytest>=8.0
pytest-django>=4.8
factory-boy>=3.3
drf-spectacular>=0.27
python-decouple>=3.8
django-filter>=23.5